Move ItemCustomizeManager to the same package as GuiItemCustomize and ItemCustomizationUtills
